+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### v0.2.5
-- New feature: modify plant grow time and space requirement
+- New feature: modify plant grow time and space requirement, option to keep plants which can't grow alive
 - Changed taming progress message to be displayed in world
 
 ### v0.2.4